シリアルポートの初期化(正しく動いてない)

+++ kernel/asmfunc.h
@@ -3,7 +3,7 @@
 
 #include "int.h"
 
-void out8(uint8 addr, uint8 data);
-uint8 in8(uint8 addr);
+void out8(uint16 addr, uint8 data);
+uint8 in8(uint16 addr);
 
 #endif
